API_building/requirements.txt
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddleocr==2.6.1.3
paddlepaddle==2.4.2
onnxruntime
pillow==10.0.1
flask==2.2.5
flask_cors==4.0.0
numpy==1.26.0
ultralytics==8.0.157
vietocr==0.3.11
opencv-python==4.6.0.6
matplotlib==3.5.3
pdf2image==1.16.3